requirements.txt
psycopg2-binary==2.9.3
loremipsum==1.0.5
